requirements.txt
discord==2.3.2
Flask==3.0.3
discord.py==2.4.0
pytest==7.2.2
Werkzeug==3.0.3
gunicorn==20.1.0
google-generativeai==0.7.2